Apply calibration output plot refactor to ENGINX scripts RE #30048
